Add static_column kwarg to CopyMapping for one-off values not in source file
This allows for fields in the model but not in the source file
to be populated with a one-off field.
An example use case is to track the originator of data. Instead
of the incoming data needing a duplicate field in every row
the copy process can include something like
static_columns={'originator':some_local_lookup_for_originator()}